Add support for loading custom tests
I have updated the use of ClasspathHelper to use the forManifest method
rather than forClassloader, as this respects the class path specified in
the jar's manifest file. Additionally, the pom file now specifies a
classpath entry for 'tests.jar'. Finally, the buildForkedCmdLine method
has been updated to include the newly added options.

In order to use this, build some tests implementing the
OneActorOneObserverTest or TwoActorsOneArbiterTest interface and package
these in a jar. Rename this jar to 'tests.jar' and place it alongside
the concurrency-torture.jar, then specify the packages containing your
test descriptions and classes with the -package and -desc command line
arguments.